src/nautilus-server/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
cryptography>=41.0.0
requests>=2.31.0
PyYAML>=6.0
cbor2>=5.4.0

